requirements.txt
streamlit
scikit-learn
transformers
sentence-transformers
numpy
requests
python-dotenv